+++ src/main/java/com/example/newhomework/service/impl/BookServiceImpl.java
@@ -2,17 +2,13 @@ package com.example.newhomework.service.impl;
 
 import com.example.newhomework.dao.impl.BookDaoImpl;
 import com.example.newhomework.dto.BookDto;
-import com.example.newhomework.entity.Author;
 import com.example.newhomework.entity.Book;
-import com.example.newhomework.entity.Licence;
 import com.example.newhomework.mapper.impl.BookToBookDtoMapperImpl;
 import com.example.newhomework.service.BookService;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
 import java.io.IOException;
-import java.io.PrintWriter;
+import java.sql.SQLException;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 
@@ -30,55 +26,22 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public int create(Book book) throws IOException {
+    public int create(Book book) throws IOException, SQLException {
         return bookDao.create(book);
-//        String name = req.getParameter("name");
-//        long author_id = Long.parseLong(req.getParameter("author_id"));
-//
-//        Book book = new Book();
-//        Author author = new Author();
-//        author.setId(author_id);
-//
-//        book.setName(name);
-//        book.setAuthor(author);
-//        PrintWriter out = resp.getWriter();
-//        if (bookDao.create(book) > 0) {
-//            out.println("запись создалась");
-//            resp.setStatus(HttpServletResponse.SC_CREATED);
-//        } else {
-//            out.println("запись не создалась");
-//        }
-
     }
 
     @Override
-    public int update(Book book) throws IOException {
-        return bookDao.update(book);
-//        Long id = Long.parseLong(req.getParameter("id"));
-//        String name = req.getParameter("name");
-//        long authorId = Long.parseLong(req.getParameter("author_id"));
-//        Book book = new Book()
-//                .setId(id)
-//                .setName(name);
-//        PrintWriter out = resp.getWriter();
-//        if (bookDao.update(book, authorId) > 0) {
-//            out.println("запись обновлена");
-//        } else {
-//            out.println("запись не обновлена");
-//        }
+    public int update(Book book) throws IOException, SQLException {
+        if (bookDao.existById(book.getId())) {
+            return bookDao.update(book);
+        } else throw new NoSuchElementException("Не найден");
     }
 
     @Override
-    public int delete(Long id) throws IOException {
-        return bookDao.delete(id);
-//        long id = Long.parseLong(req.getParameter("id"));
-//        PrintWriter out = resp.getWriter();
-//        if (bookDao.delete(id) > 0) {
-//            out.println("запись удалена");
-//            resp.setStatus(HttpServletResponse.SC_OK);
-//        } else {
-//            out.println("запись не удалена");
-//        }
+    public int delete(Long id) throws IOException, SQLException {
+        if (bookDao.existById(id)) {
+            return bookDao.delete(id);
+        } else throw new NoSuchElementException("Не найден");
     }
